requirements.txt
requests
celery-with-redis
flower
redis
lxml
simplejson
pyeurofx
networkx
